server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-telegram-bot>=20.7
tapo-py3>=0.3.0
wakeonlan>=3.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
